isolated nat scalar tests
